+++ oanda/go.mod
@@ -1,0 +1,5 @@
+module github.com/davidhintelmann/Oanda-Go/oanda
+
+go 1.20
+
+require github.com/davidhintelmann/Oanda-Go/restful v0.0.0-20240917040452-b3d836644c9a
