--- restful/go.mod
@@ -1,3 +1,0 @@
-module github.com/davidhintelmann/Oanda-Go/restful
-
-go 1.20
